+++ source/release-notes.md
@@ -800,3 +800,4 @@
 
 27.11.2018
 * add option to ignore job execution limit warning alerts
+* fix issue: There is already an object named '##tdp_sql_version_requests' in the database