+++ source/release-notes.md
@@ -801,3 +801,8 @@
 27.11.2018
 * add option to ignore job execution limit warning alerts
 * fix issue: There is already an object named '##tdp_sql_version_requests' in the database
+
+------------
+#### December 2018 new features & fixes
+10.12.2018
+* monitoring: add support for replication monitoring when distributor is on a remote instance